add new directive `ready-to-go-stacks`
